requirements.txt
pandas
numpy
matplotlib
seaborn
plotly
requests
python-dotenv